+++ lib/src/cli/commands/diff_command.dart
@@ -11,6 +11,8 @@ import 'command_mixin.dart';
 
 String _optionNameOld = 'old';
 String _optionNameNew = 'new';
+String _optionNameOldCopyDepth = 'old-copy-depth';
+String _optionNameNewCopyDepth = 'new-copy-depth';
 String _optionNameCheckVersions = 'check-versions';
 String _optionNameIgnorePrerelease = 'ignore-prerelease';
 String _optionNameNoMergeBaseClasses = 'no-merge-base-classes';
@@ -37,11 +39,19 @@ class DiffCommand extends Command<int> with CommandMixin {
       mandatory: true,
       help: 'New package reference. $packageRefExplanation',
     );
+    argParser.addOption(
+      _optionNameOldCopyDepth,
+      help: copyDepthExplanation('old'),
+    );
+    argParser.addOption(
+      _optionNameNewCopyDepth,
+      help: copyDepthExplanation('new'),
+    );
     argParser.addFlag(
       _optionNameCheckVersions,
       help: '''
 Determines if the version of the new package should be checked.
-Takes the changes of the diff and checks if the new version follows semver. 
+Takes the changes of the diff and checks if the new version follows semver.
 Influences tool return value.
 ''',
       defaultsTo: true,
@@ -56,10 +66,10 @@ Influences tool return value.
     argParser.addFlag(
       _optionNameIgnorePrerelease,
       help: '''
-Determines if the pre-release aspect of the version 
-shall be ignored when checking versions. 
+Determines if the pre-release aspect of the version
+shall be ignored when checking versions.
 This only makes sense in combination with --$_optionNameCheckVersions.
-You may want to do this if you want to make sure 
+You may want to do this if you want to make sure
 (in your CI) that the version - once ready - matches semver.
 ''',
       defaultsTo: false,
@@ -89,6 +99,14 @@ You may want to do this if you want to make sure
   Future<int> run() async {
     final oldPackageRef = PackageRef(argResults![_optionNameOld]);
     final newPackageRef = PackageRef(argResults![_optionNameNew]);
+    final oldPackageCopyDepthString = argResults![_optionNameOldCopyDepth];
+    final oldPackageCopyDepth = oldPackageCopyDepthString != null
+        ? int.parse(oldPackageCopyDepthString)
+        : null;
+    final newPackageCopyDepthString = argResults![_optionNameNewCopyDepth];
+    final newPackageCopyDepth = newPackageCopyDepthString != null
+        ? int.parse(newPackageCopyDepthString)
+        : null;
     final checkVersions = argResults![_optionNameCheckVersions] as bool;
     final ignorePrerelease = argResults![_optionNameIgnorePrerelease] as bool;
     final doCheckSdkVersion = argResults![_optionNameCheckSdkVersion] as bool;
@@ -100,8 +118,10 @@ You may want to do this if you want to make sure
         (element) =>
             element.name == argResults![_optionNameDependencyCheckMode]);
 
-    final preparedOldPackageRef = await prepare(oldPackageRef);
-    final preparedNewPackageRef = await prepare(newPackageRef);
+    final preparedOldPackageRef =
+        await prepare(oldPackageRef, depth: oldPackageCopyDepth);
+    final preparedNewPackageRef =
+        await prepare(newPackageRef, depth: newPackageCopyDepth);
 
     final oldPackageApi = await analyze(
       preparedOldPackageRef,