+++ lib/src/cli/commands/extract_command.dart
@@ -7,6 +7,7 @@ import '../package_ref.dart';
 import 'command_mixin.dart';
 
 String _optionNameInput = 'input';
+String _optionNameInputCopyDepth = 'input-copy-depth';
 String _optionNameOutput = 'output';
 String _optionNameNoMergeBaseClasses = 'no-merge-base-classes';
 String _optionNameNoAnalyzePlatformConstraints =
@@ -26,6 +27,10 @@ class ExtractCommand extends Command<int> with CommandMixin {
       _optionNameInput,
       help: 'Input package ref. $packageRefExplanation',
       mandatory: true,
+    );
+    argParser.addOption(
+      _optionNameInputCopyDepth,
+      help: copyDepthExplanation('input'),
     );
     argParser.addOption(
       _optionNameOutput,
@@ -51,12 +56,17 @@ If not specified the extracted API will be printed to the console.
   @override
   Future<int> run() async {
     final packageRef = PackageRef(argResults![_optionNameInput]);
+    final packageCopyDepthString = argResults![_optionNameInputCopyDepth];
+    final packageCopyDepth = packageCopyDepthString != null
+        ? int.parse(packageCopyDepthString)
+        : null;
     final noMergeBaseClasses =
         argResults![_optionNameNoMergeBaseClasses] as bool;
     final noAnalyzePlatformConstraints =
         argResults![_optionNameNoAnalyzePlatformConstraints] as bool;
 
-    final preparedPackageRef = await prepare(packageRef);
+    final preparedPackageRef =
+        await prepare(packageRef, depth: packageCopyDepth);
     final packageApi = await analyze(
       preparedPackageRef,
       doMergeBaseClasses: !noMergeBaseClasses,